Face-Recognition-Attendance-System-new - Copy/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
Pillow>=8.0.0
pandas>=1.2.0
requests>=2.25.0 